+++ Source/Scene/CloudCollection.js
@@ -3,6 +3,7 @@ import Buffer from "../Renderer/Buffer.js";
 import BufferUsage from "../Renderer/BufferUsage.js";
 import Cartesian3 from "../Core/Cartesian3.js";
 import Check from "../Core/Check.js";
+import Color from "../Core/Color.js";
 import ComputeCommand from "../Renderer/ComputeCommand.js";
 import CloudType from "./CloudType.js";
 import CloudCollectionFS from "../Shaders/CloudCollectionFS.js";
@@ -40,6 +41,8 @@ var attributeLocationsBatched = {
   positionLowAndScaleY: 1,
   packedAttribute0: 2, // show, brightness, direction
   packedAttribute1: 3, // cloudSize, slice
+  packedAttribute2: 4, // cloudColor
+
 };
 
 var attributeLocationsInstanced = {
@@ -48,6 +51,7 @@ var attributeLocationsInstanced = {
   positionLowAndScaleY: 2,
   packedAttribute0: 3, // show, brightness
   packedAttribute1: 4, // cloudSize, slice
+  packedAttribute2: 5, // cloudColor
 };
 
 var SHOW_INDEX = CumulusCloud.SHOW_INDEX;
@@ -57,6 +61,8 @@ var MAXIMUM_SIZE_INDEX = CumulusCloud.MAXIMUM_SIZE_INDEX;
 var SLICE_INDEX = CumulusCloud.SLICE_INDEX;
 var BRIGHTNESS_INDEX = CumulusCloud.BRIGHTNESS_INDEX;
 var NUMBER_OF_PROPERTIES = CumulusCloud.NUMBER_OF_PROPERTIES;
+var COLOR_INDEX = CumulusCloud.COLOR_INDEX;
+
 
 /**
  * A renderable collection of clouds in the 3D scene.
@@ -163,6 +169,9 @@ function CloudCollection(options) {
    */
   this.noiseOffset = Cartesian3.clone(
     defaultValue(options.noiseOffset, Cartesian3.ZERO)
+  );
+  this.color = Cartesian3.clone(
+    defaultValue(options.color, Color.WHITE)
   );
 
   this._loading = false;
@@ -569,7 +578,12 @@ function createVAF(context, numberOfClouds, instanced) {
       componentsPerAttribute: 4,
       componentDatatype: ComponentDatatype.FLOAT,
       usage: BufferUsage.STATIC_DRAW,
-    },
+    },{
+      index: attributeLocations.packedAttribute2,
+      componentsPerAttribute: 4,
+      componentDatatype: ComponentDatatype.FLOAT,
+      usage: BufferUsage.STATIC_DRAW,
+    }
   ];
 
   if (instanced) {
@@ -653,10 +667,27 @@ function writePackedAttribute1(cloudCollection, frameState, vafWriters, cloud) {
   }
 }
 
+function writePackedAttribute2(cloudCollection, frameState, vafWriters, cloud) {
+  var i;
+  var writer = vafWriters[attributeLocations.packedAttribute2];
+  var color = cloud.color;
+  var slice = cloud.slice;
+  if (cloudCollection._instanced) {
+    i = cloud._index;
+    writer(i, color.red, color.green, color.blue, slice);
+  } else {
+    i = cloud._index * 4;
+    writer(i + 0, color.red, color.green, color.blue, slice);
+    writer(i + 1, color.red, color.green, color.blue, slice);
+    writer(i + 2, color.red, color.green, color.blue, slice);
+    writer(i + 3, color.red, color.green, color.blue, slice);
+  }
+}
 function writeCloud(cloudCollection, frameState, vafWriters, cloud) {
   writePositionAndScale(cloudCollection, frameState, vafWriters, cloud);
   writePackedAttribute0(cloudCollection, frameState, vafWriters, cloud);
   writePackedAttribute1(cloudCollection, frameState, vafWriters, cloud);
+  writePackedAttribute2(cloudCollection, frameState, vafWriters, cloud);
 }
 
 function createNoiseTexture(cloudCollection, frameState, vsSource, fsSource) {
@@ -766,6 +797,10 @@ function updateClouds(cloudCollection, frameState) {
 
   if (properties[MAXIMUM_SIZE_INDEX] || properties[SLICE_INDEX]) {
     writers.push(writePackedAttribute1);
+  }
+
+  if (properties[COLOR_INDEX] || properties[SLICE_INDEX]) {
+    writers.push(writePackedAttribute2);
   }
 
   var numWriters = writers.length;