+++ Source/Shaders/CloudCollectionVS.glsl
@@ -5,9 +5,11 @@ attribute vec4 positionHighAndScaleX;
 attribute vec4 positionLowAndScaleY;
 attribute vec4 packedAttribute0;
 attribute vec4 packedAttribute1;
+attribute vec4 packedAttribute2;
 
 varying vec2 v_offset;
 varying vec3 v_maximumSize;
+varying vec3 v_color;
 varying float v_slice;
 varying float v_brightness;
 
@@ -21,6 +23,7 @@ void main() {
     float brightness = packedAttribute0.y;
     vec2 coordinates = packedAttribute0.wz;
     vec3 maximumSize = packedAttribute1.xyz;
+    vec3 color = packedAttribute2.xyz;
     float slice = packedAttribute1.w;
 
 #ifdef INSTANCED
@@ -40,6 +43,7 @@ void main() {
 
     v_offset = offset;
     v_maximumSize = maximumSize;
+    v_color=color;
     v_slice = slice;
     v_brightness = brightness;
 }