+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - Fixed zip.js configurations causing CesiumJS to not work with Node 16. [#9861](https://github.com/CesiumGS/cesium/pull/9861)
 - Fixed a bug in `Rectangle.union` with rectangles that span the entire globe. [#9866](https://github.com/CesiumGS/cesium/pull/9866)
+- Added CumulusCloud.color ,CumulusCloud can customize colors
 
 ### 1.86 - 2021-10-01